spooky-core: move new style of Bevy logic into the crate
