+++ desktop/Cargo.toml
@@ -11,7 +11,7 @@ bevy = "0.16.0-rc.2"
 rand = "0.9.0"
 rand_chacha = "0.9.0"
 maze_generator = { git = "https://github.com/georgik/maze_generator", branch = "feature/no_std", default-features = false, features = [ "recursive_backtracking"], optional = true }
-
+spooky-core = { path = "../spooky-core", features = ["std"] }
 
 [features]
 default = [ "static_maze" ]
