+++ spooky-core/src/maze.rs
@@ -1,3 +1,5 @@
+// If you want dynamic maze generation, enable the "dynamic_maze" feature
+// and ensure the dependency on `maze_generator` is added to Cargo.toml.
 #[cfg(feature = "dynamic_maze")]
 use maze_generator::{prelude::*, recursive_backtracking::RbGenerator};
 
@@ -11,8 +13,8 @@ pub struct Coin {
 }
 
 impl Coin {
-    pub fn new(x: i32, y: i32) -> Coin {
-        Coin { x, y }
+    pub fn new(x: i32, y: i32) -> Self {
+        Self { x, y }
     }
 }
 
@@ -24,6 +26,7 @@ pub struct Npc {
     pub vector_y: i32,
 }
 
+#[derive(Clone)]
 pub struct Maze {
     pub width: u32,
     pub height: u32,
@@ -35,7 +38,6 @@ pub struct Maze {
     pub npcs: [Npc; 5],
     pub walkers: [Coin; 5],
     pub dynamites: [Coin; 1],
-    // Tile map should have small border top line and left column
     pub offset: u32,
     pub tile_width: u32,
     pub tile_height: u32,
@@ -43,8 +45,10 @@ pub struct Maze {
 }
 
 impl Maze {
-    pub fn new(width: u32, height: u32, seed: Option<[u8; 32]>) -> Maze {
-        Maze {
+    pub const MARGIN: i32 = 10;
+    /// Create a new maze with the given dimensions and an optional seed.
+    pub fn new(width: u32, height: u32, seed: Option<[u8; 32]>) -> Self {
+        Self {
             width,
             height,
             visible_width: 21,
@@ -216,9 +220,8 @@ impl Maze {
             walkers: [Coin { x: -1, y: -1 }; 5],
             dynamites: [Coin { x: -1, y: -1 }; 1],
             rng: match seed {
-                // None => ChaChaRng::from_entropy(), - from_entropy is not present in latest rand
                 None => ChaChaRng::from_seed([42; 32]),
-                Some(seed) => ChaChaRng::from_seed(seed),
+                Some(s) => ChaChaRng::from_seed(s),
             },
         }
     }
@@ -228,43 +231,63 @@ impl Maze {
     }
 
     pub fn check_boundary_collision(&self, x: i32, y: i32) -> bool {
-        if x < 0
+        x < 0
             || y < 0
             || x >= (self.width * self.tile_width) as i32
             || y >= (self.height * self.tile_height) as i32
-        {
-            return true;
-        }
-        false
     }
 
     pub fn check_wall_collision(&self, x: i32, y: i32) -> bool {
-        if x < 0
-            || y < 0
-            || x >= (self.width * self.tile_width) as i32
-            || y >= (self.height * self.tile_height) as i32
-        {
+        let (left, bottom, right, top) = self.playable_bounds();
+        // Outside the playable area is considered a collision.
+        if x < left || x >= right || y < bottom || y >= top {
             return true;
         }
-        let tile_x = x / self.tile_width as i32;
-        let tile_y = y / self.tile_height as i32;
-        let tile_index = (tile_y * self.width as i32 + tile_x) as usize;
+        // Compute tile coordinates relative to the playable area.
+        let tile_x = (x - left) / self.tile_width as i32;
+        let tile_y = (y - bottom) / self.tile_height as i32;
+        // Convert world tile_y (0 at bottom) to maze row index (0 at top).
+        let maze_row = (self.height as i32 - 1) - tile_y;
+        if tile_x < 0 || maze_row < 0 || tile_x >= self.width as i32 || maze_row >= self.height as i32 {
+            return true;
+        }
+        let tile_index = (maze_row * self.width as i32 + tile_x) as usize;
         self.data[tile_index] == 1
     }
 
-    pub fn get_random_coordinates(&mut self) -> (i32, i32) {
-        let mut x = (self.get_rand() % (self.width as i32 - 2) + 1) * self.tile_width as i32;
-        let mut y = (self.get_rand() % (self.height as i32 - 2) + 1) * self.tile_height as i32;
-        while self.check_wall_collision(x, y) {
-            x = (self.get_rand() % (self.width as i32 - 2) + 1) * self.tile_width as i32;
-            y = (self.get_rand() % (self.height as i32 - 2) + 1) * self.tile_height as i32;
+    pub fn valid_coordinates(&self) -> Vec<(i32, i32)> {
+        let mut coords = Vec::new();
+        for ty in 0..self.height as i32 {
+            for tx in 0..self.width as i32 {
+                // Compute pixel coordinates (e.g., center of the tile)
+                let x = tx * self.tile_width as i32 + (self.tile_width as i32 / 2);
+                let y = ty * self.tile_height as i32 + (self.tile_height as i32 / 2);
+                // Check that this tile is walkable (i.e. not a wall)
+                let tile_index = (ty * self.width as i32 + tx) as usize;
+                if self.data[tile_index] == 0 {
+                    coords.push((x, y));
+                }
+            }
         }
-        (x, y)
+        coords
+    }
+
+    pub fn get_random_coordinates(&mut self) -> (i32, i32) {
+        let valid = self.valid_coordinates();
+        if valid.is_empty() {
+            // Fallback coordinate if no valid tile exists.
+            (1, 1)
+        } else {
+            let idx = self.get_rand() as usize % valid.len();
+            valid[idx]
+        }
     }
 
     pub fn generate_coins(&mut self) {
         for index in 0..100 {
-            (self.coins[index].x, self.coins[index].y) = self.get_random_coordinates();
+            let (new_x, new_y) = self.get_random_coordinates();
+            self.coins[index].x = new_x;
+            self.coins[index].y = new_y;
         }
         self.coin_counter = 100;
     }
@@ -273,7 +296,9 @@ impl Maze {
         let mut relocate_counter = 0;
         for index in 0..100 {
             if self.coins[index].x == -1 && self.coins[index].y == -1 {
-                (self.coins[index].x, self.coins[index].y) = self.get_random_coordinates();
+                let (new_x, new_y) = self.get_random_coordinates();
+                self.coins[index].x = new_x;
+                self.coins[index].y = new_y;
                 relocate_counter += 1;
                 self.coin_counter += 1;
                 if relocate_counter == amount {
@@ -285,58 +310,44 @@ impl Maze {
 
     pub fn generate_walkers(&mut self) {
         for index in 0..5 {
-            (self.walkers[index].x, self.walkers[index].y) = self.get_random_coordinates();
+            let (new_x, new_y) = self.get_random_coordinates();
+            self.walkers[index].x = new_x;
+            self.walkers[index].y = new_y;
         }
     }
 
     pub fn generate_dynamites(&mut self) {
         for index in 0..1 {
-            (self.dynamites[index].x, self.dynamites[index].y) = self.get_random_coordinates();
+            let (new_x, new_y) = self.get_random_coordinates();
+            self.dynamites[index].x = new_x;
+            self.dynamites[index].y = new_y;
         }
     }
 
     pub fn generate_npcs(&mut self) {
         for index in 0..5 {
-            (self.npcs[index].x, self.npcs[index].y) = self.get_random_coordinates();
+            let (new_x, new_y) = self.get_random_coordinates();
+            self.npcs[index].x = new_x;
+            self.npcs[index].y = new_y;
             self.npcs[index].vector_x = 1;
             self.npcs[index].vector_y = 1;
         }
     }
 
     pub fn get_coin_at(&self, x: i32, y: i32) -> Option<Coin> {
-        for coin in self.coins.iter() {
-            if coin.x == x && coin.y == y {
-                return Some(*coin);
-            }
-        }
-        None
+        self.coins.iter().copied().find(|coin| coin.x == x && coin.y == y)
     }
 
     pub fn get_npc_at(&self, x: i32, y: i32) -> Option<Npc> {
-        for npc in self.npcs.iter() {
-            if npc.x == x && npc.y == y {
-                return Some(*npc);
-            }
-        }
-        None
+        self.npcs.iter().copied().find(|npc| npc.x == x && npc.y == y)
     }
 
     pub fn get_walker_at(&self, x: i32, y: i32) -> Option<Coin> {
-        for walker in self.walkers.iter() {
-            if walker.x == x && walker.y == y {
-                return Some(*walker);
-            }
-        }
-        None
+        self.walkers.iter().copied().find(|walker| walker.x == x && walker.y == y)
     }
 
     pub fn get_dynamite_at(&self, x: i32, y: i32) -> Option<Coin> {
-        for dynamite in self.dynamites.iter() {
-            if dynamite.x == x && dynamite.y == y {
-                return Some(*dynamite);
-            }
-        }
-        None
+        self.dynamites.iter().copied().find(|d| d.x == x && d.y == y)
     }
 
     pub fn remove_coin(&mut self, coin: Coin) {
@@ -354,7 +365,9 @@ impl Maze {
     pub fn relocate_walker(&mut self, walker: Coin) {
         for index in 0..5 {
             if self.walkers[index].x == walker.x && self.walkers[index].y == walker.y {
-                (self.walkers[index].x, self.walkers[index].y) = self.get_random_coordinates();
+                let (new_x, new_y) = self.get_random_coordinates();
+                self.walkers[index].x = new_x;
+                self.walkers[index].y = new_y;
             }
         }
     }
@@ -362,7 +375,9 @@ impl Maze {
     pub fn relocate_dynamite(&mut self, dynamite: Coin) {
         for index in 0..1 {
             if self.dynamites[index].x == dynamite.x && self.dynamites[index].y == dynamite.y {
-                (self.dynamites[index].x, self.dynamites[index].y) = self.get_random_coordinates();
+                let (new_x, new_y) = self.get_random_coordinates();
+                self.dynamites[index].x = new_x;
+                self.dynamites[index].y = new_y;
             }
         }
     }
@@ -370,26 +385,23 @@ impl Maze {
     pub fn set_tile_at(&mut self, x: i32, y: i32, tile: u8) {
         let tile_x = x / self.tile_width as i32;
         let tile_y = y / self.tile_height as i32;
-        let tile_index = (tile_y * self.width as i32 + tile_x) as usize;
-
         if tile_x < 0 || tile_y < 0 || tile_x >= self.width as i32 || tile_y >= self.height as i32 {
             return;
         }
-
+        let tile_index = (tile_y * self.width as i32 + tile_x) as usize;
         self.data[tile_index] = tile;
     }
 
     pub fn place_dynamite(&mut self, x: i32, y: i32) {
-        self.set_tile_at(x - self.tile_width as i32, y - self.tile_height as i32, 2);
-        self.set_tile_at(x, y - self.tile_height as i32, 2);
-        self.set_tile_at(x + self.tile_width as i32, y - self.tile_height as i32, 2);
-
-        self.set_tile_at(x - self.tile_width as i32, y, 2);
-        self.set_tile_at(x + self.tile_width as i32, y, 2);
-
-        self.set_tile_at(x - self.tile_width as i32, y + self.tile_height as i32, 2);
-        self.set_tile_at(x, y + self.tile_height as i32, 2);
-        self.set_tile_at(x + self.tile_width as i32, y + self.tile_height as i32, 2);
+        let tw = self.tile_width as i32;
+        self.set_tile_at(x - tw, y - tw, 2);
+        self.set_tile_at(x, y - tw, 2);
+        self.set_tile_at(x + tw, y - tw, 2);
+        self.set_tile_at(x - tw, y, 2);
+        self.set_tile_at(x + tw, y, 2);
+        self.set_tile_at(x - tw, y + tw, 2);
+        self.set_tile_at(x, y + tw, 2);
+        self.set_tile_at(x + tw, y + tw, 2);
     }
 
     fn get_random_vector(&mut self) -> (i32, i32) {
@@ -405,10 +417,12 @@ impl Maze {
         for index in 0..5 {
             let mut x = self.npcs[index].x;
             let mut y = self.npcs[index].y;
-            x += self.npcs[index].vector_x * 16;
-            y += self.npcs[index].vector_y * 16;
+            x += self.npcs[index].vector_x * self.tile_width as i32;
+            y += self.npcs[index].vector_y * self.tile_height as i32;
             if self.check_wall_collision(x, y) {
-                (self.npcs[index].vector_x, self.npcs[index].vector_y) = self.get_random_vector();
+                let (vx, vy) = self.get_random_vector();
+                self.npcs[index].vector_x = vx;
+                self.npcs[index].vector_y = vy;
             } else {
                 self.npcs[index].x = x;
                 self.npcs[index].y = y;
@@ -417,43 +431,38 @@ impl Maze {
     }
 
     #[cfg(feature = "static_maze")]
-    pub fn generate_maze(&mut self, _graph_width: usize, _graph_height: usize) {}
+    pub fn generate_maze(&mut self, _graph_width: usize, _graph_height: usize) {
+        // No dynamic generation in static mode.
+    }
 
     #[cfg(feature = "dynamic_maze")]
     pub fn generate_maze(&mut self, graph_width: usize, graph_height: usize) {
-        // let mut rng = Rng::new(peripherals.RNG);
-        // let mut rng = Rng::new( 0x12345678 );
         let seed_buffer = [0u8; 32];
-        // match &self.rng {
-        //     Some(rng) => rng.fill_bytes(&mut seed_buffer),
-        //     None => {}
-        // };
-        // #[cfg(feature = "getrandom")]
-        // getrandom::getrandom(&mut seed_buffer).unwrap();
-
         let mut generator = RbGenerator::new(Some(seed_buffer));
-        let maze_graph = generator
-            .generate(graph_width as i32, graph_height as i32)
-            .unwrap();
-
-        for y in 1usize..graph_height {
-            for x in 1usize..graph_width {
+        let maze_graph = generator.generate(graph_width as i32, graph_height as i32).unwrap();
+        for y in 1..graph_height {
+            for x in 1..graph_width {
                 let field = maze_graph
-                    .get_field(&(x.try_into().unwrap(), y.try_into().unwrap()).into())
+                    .get_field(&(x as i32, y as i32).into())
                     .unwrap();
-                let tile_index =
-                    (x - 1) * 2 + (y - 1) * 2 * (self.width as usize) + (self.offset as usize);
-
+                let tile_index = (x - 1) * 2 + (y - 1) * 2 * (self.width as usize) + (self.offset as usize);
                 self.data[tile_index] = 0;
-
                 if field.has_passage(&Direction::West) {
                     self.data[tile_index + 1] = 0;
                 }
-
                 if field.has_passage(&Direction::South) {
                     self.data[tile_index + (self.width as usize)] = 0;
                 }
             }
         }
+    }
+
+    pub fn playable_bounds(&self) -> (i32, i32, i32, i32) {
+        let margin = Self::MARGIN;
+        let left = margin * self.tile_width as i32;
+        let bottom = margin * self.tile_height as i32;
+        let right = left + self.width as i32 * self.tile_width as i32;
+        let top = bottom + self.height as i32 * self.tile_height as i32;
+        (left, bottom, right, top)
     }
 }
