--- spooky-core/src/movement_controller.rs
@@ -1,7 +1,0 @@
-use crate::engine::Action;
-
-pub trait MovementController {
-    fn set_active(&mut self, index: usize);
-    fn tick(&mut self);
-    fn get_movement(&self) -> Action;
-}
