+++ spooky-core/src/resources.rs
