+++ spooky-core/src/systems/game_logic.rs
