+++ spooky-core/src/systems/setup.rs
